comments

+++ calendar/php/deleteEvent.php
@@ -3,25 +3,25 @@
     ini_set("session.cookie_httponly", 1);
 
     session_start();
-    session_name('logged in');
 
     header("Content-Type: application/json");
+
+    //Check to make sure HTTP User Agent is consistent
+    $previous_ua = @$_SESSION['useragent'];
+    $current_ua = $_SERVER['HTTP_USER_AGENT'];
+
+    if(isset($_SESSION['useragent']) && $previous_ua !== $current_ua){
+        echo json_encode(array("Session hijack detected" => true));
+        die("Session hijack detected");
+    }else{
+        $_SESSION['useragent'] = $current_ua;
+    }
 
     //Check to make sure user is not already logged in
     if(!isset($_SESSION['user'])) {
         echo json_encode(array("nosession" => true));
         die();
     }
-
-    $previous_ua = @$_SESSION['useragent'];
-    $current_ua = $_SERVER['HTTP_USER_AGENT'];
-
-    if(isset($_SESSION['useragent']) && $previous_ua !== $current_ua){
-        die("Session hijack detected");
-    }else{
-        $_SESSION['useragent'] = $current_ua;
-    }
-
 
     $arr = array();
 
@@ -30,12 +30,23 @@
     $json_obj = json_decode($json_str, true);
 
     $id = $json_obj['id'];
+    $token = $json_obj['token'];
 
-    
+    //Get token if exists
+    if($json_obj['token'] != '') {
+        $token = $json_obj['token'];
+    }
+
+    //Verify token to prevent CSRF attack
+    if($token != $_SESSION['token']) {
+        echo json_encode(array("Invalid Token" => true));
+        die();
+    } else { 
+        $arr['token'] = true;
+    }
+
     //connect to database with wustl user
     $conn = new mysqli('localhost', 'wustl_inst', 'wustl_pass', 'calendar');
-
-
 
     //exit if connection fails
     if ($conn->connect_error) {
@@ -44,9 +55,9 @@
         die();
     }
 
-
-    $query = $conn->prepare('DELETE FROM events WHERE id=(?)');
-    $query->bind_param("i", $id);
+    //Delete event from database
+    $query = $conn->prepare('DELETE FROM events WHERE id=(?) AND userId=(?)');
+    $query->bind_param("ii", $id, $_SESSION['userid']);
     if($query->execute()) {
         $arr['success'] = true;
     } else {
@@ -54,5 +65,4 @@
     }
 
     echo json_encode($arr);
-
 ?>
